pyproject.toml
[project]
name = "graph-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.125.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-neo4j>=0.6.0",
    "langchain-openai>=1.1.6",
    "neo4j<6.0.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
